Torch.Data.CsvDataset

Documentation

tsvDataset :: forall (isNamed :: NamedColumns) batches. FilePath -> CsvDataset' batches isNamed

csvDataset :: forall (isNamed :: NamedColumns) batches. FilePath -> CsvDataset' batches isNamed

type CsvDataset batches = CsvDataset' batches Unnamed

Use if you want to decode with a record that has a FromRecord instance.

type CsvDatasetNamed batches = CsvDataset' batches Named

Use CsvDatasetNamed if you want to decode with a record that has FromNamedRecord instance, | decoding fields of a given name from a csv file,

data CsvDataset' batches (named :: NamedColumns)

Constructors

CsvDataset'

Fields

filePath :: FilePath

delimiter :: !Word8

hasHeader :: HasHeader

batchSize :: Int

shuffle :: Maybe BufferSize

dropLast :: Bool

class FromField a where

A type that can be converted from a single CSV field, with the possibility of failure.

When writing an instance, use empty, mzero, or fail to make a conversion fail, e.g. if a Field can't be converted to the given type.

Example type and instance:

{-# LANGUAGE OverloadedStrings #-}

data Color = Red | Green | Blue

instance FromField Color where
    parseField s
        | s == "R"  = pure Red
        | s == "G"  = pure Green
        | s == "B"  = pure Blue
        | otherwise = mzero

Methods

parseField :: Field -> Parser a

class FromRecord a where

A type that can be converted from a single CSV record, with the possibility of failure.

When writing an instance, use empty, mzero, or fail to make a conversion fail, e.g. if a Record has the wrong number of columns.

Given this example data:

John,56
Jane,55

here's an example type and instance:

data Person = Person { name :: !Text, age :: !Int }

instance FromRecord Person where
    parseRecord v
        | length v == 2 = Person <$>
                          v .! 0 <*>
                          v .! 1
        | otherwise     = mzero

Minimal complete definition

Nothing

Methods

parseRecord :: Record -> Parser a

class FromNamedRecord a where

A type that can be converted from a single CSV record, with the possibility of failure.

When writing an instance, use empty, mzero, or fail to make a conversion fail, e.g. if a Record has the wrong number of columns.

Given this example data:

name,age
John,56
Jane,55

here's an example type and instance:

{-# LANGUAGE OverloadedStrings #-}

data Person = Person { name :: !Text, age :: !Int }

instance FromNamedRecord Person where
    parseNamedRecord m = Person <$>
                         m .: "name" <*>
                         m .: "age"

Note the use of the OverloadedStrings language extension which enables ByteString values to be written as string literals.

Minimal complete definition

Nothing

Methods

parseNamedRecord :: NamedRecord -> Parser a