Torch.Data.CsvDatastream

Documentation

type BufferSize = Int

data NamedColumns

Constructors

Unnamed

Named

data CsvDatastream' batches (named :: NamedColumns)

A CSV datastream. The datastream instance of this type streams samples of batches from a CSV file at the specified file path. Batches are yielded in constant memory, but if shuffling is enabled, then there will be at most BufferSize records stored in memory.

Constructors

CsvDatastream'

Fields

filePath :: FilePath

CSV file path.

delimiter :: !Word8

Column delimiter.

hasHeader :: HasHeader

Does the file have a header?

batchSize :: Int

Batch size. , filter :: Maybe (batches -> Bool)

bufferedShuffle :: Maybe BufferSize

Buffered shuffle with specified buffer size.

dropLast :: Bool

Drop the last batch if it is less than batch size.

type CsvDatastream batches = CsvDatastream' batches Unnamed

A specialized version of CsvDatastream'. Use this type if you want to decode a CSV file with records defined by the order of the columns.

type CsvDatastreamNamed batches = CsvDatastream' batches Named

A specialized version of CsvDatastream'. Use this type if you want to decode a CSV file with records that have FromNamedRecord instance. This decodes each field of the record by the corresponding column with the given header name.

csvDatastream :: forall (isNamed :: NamedColumns) batches. FilePath -> CsvDatastream' batches isNamed

Produce a CsvDatastream' from the given file with default options, and comma separated columns.

tsvDatastream :: forall (isNamed :: NamedColumns) batches. FilePath -> CsvDatastream' batches isNamed

Produce a CsvDatastream' from the given file with default options, and tab separated columns.

Reexports

class FromField a where

A type that can be converted from a single CSV field, with the possibility of failure.

When writing an instance, use empty, mzero, or fail to make a conversion fail, e.g. if a Field can't be converted to the given type.

Example type and instance:

{-# LANGUAGE OverloadedStrings #-}

data Color = Red | Green | Blue

instance FromField Color where
    parseField s
        | s == "R"  = pure Red
        | s == "G"  = pure Green
        | s == "B"  = pure Blue
        | otherwise = mzero

Methods

parseField :: Field -> Parser a

class FromRecord a where

A type that can be converted from a single CSV record, with the possibility of failure.

When writing an instance, use empty, mzero, or fail to make a conversion fail, e.g. if a Record has the wrong number of columns.

Given this example data:

John,56
Jane,55

here's an example type and instance:

data Person = Person { name :: !Text, age :: !Int }

instance FromRecord Person where
    parseRecord v
        | length v == 2 = Person <$>
                          v .! 0 <*>
                          v .! 1
        | otherwise     = mzero

Minimal complete definition

Nothing

Methods

parseRecord :: Record -> Parser a

class FromNamedRecord a where

A type that can be converted from a single CSV record, with the possibility of failure.

When writing an instance, use empty, mzero, or fail to make a conversion fail, e.g. if a Record has the wrong number of columns.

Given this example data:

name,age
John,56
Jane,55

here's an example type and instance:

{-# LANGUAGE OverloadedStrings #-}

data Person = Person { name :: !Text, age :: !Int }

instance FromNamedRecord Person where
    parseNamedRecord m = Person <$>
                         m .: "name" <*>
                         m .: "age"

Note the use of the OverloadedStrings language extension which enables ByteString values to be written as string literals.

Minimal complete definition

Nothing

Methods

parseNamedRecord :: NamedRecord -> Parser a