Torch.Data.Dataset

Documentation

data CollatedDataset m dataset batch collatedBatch

This class is actually not very useful. | It would actually be better to define a transform | on top of another dataset, since then we can do this in parallel

Constructors

CollatedDataset

Fields

set :: dataset

chunkSize :: Int

collateFn :: Pipe [batch] collatedBatch m ()