Torch.Distributions.Distribution

Documentation

data Scale

class Distribution a where

Methods

batchShape :: a -> [Int]

eventShape :: a -> [Int]

expand :: a -> [Int] -> a

support :: a -> Constraint

mean :: a -> Tensor

variance :: a -> Tensor

sample :: a -> [Int] -> IO Tensor

logProb :: a -> Tensor -> Tensor

entropy :: a -> Tensor

enumerateSupport :: a -> Bool -> Tensor

stddev :: Distribution a => a -> Tensor

perplexity :: Distribution a => a -> Tensor

logitsToProbs :: Bool -> Tensor -> Tensor

Converts a tensor of logits into probabilities. Note that for the | binary case, each value denotes log odds, whereas for the | multi-dimensional case, the values along the last dimension denote | the log probabilities (possibly unnormalized) of the events.

clampProbs :: Tensor -> Tensor

probsToLogits :: Bool -> Tensor -> Tensor

Converts a tensor of probabilities into logits. For the binary case, | this denotes the probability of occurrence of the event indexed by `1`. | For the multi-dimensional case, the values along the last dimension | denote the probabilities of occurrence of each of the events.

extendedShape :: Distribution a => a -> [Int] -> [Int]

Returns the size of the sample returned by the distribution, given | a sampleShape. Note, that the batch and event shapes of a distribution | instance are fixed at the time of construction. If this is empty, the | returned shape is upcast to (1,). | Args: | sampleShape (torch.Size): the size of the sample to be drawn.