Torch.HList

Documentation

type family ListLength (xs :: [k]) :: Nat where ...

Equations

ListLength '[] = 0

ListLength (_h ': t) = 1 + ListLength t

data family HList (xs :: [k])

pattern (:.) :: forall x (xs :: [Type]). x -> HList xs -> HList (x ': xs) infixr 2

class Apply f a b where

Methods

apply :: f -> a -> b

class Apply' f a b | f a -> b where

Stronger version of Apply that allows for better inference of the return type

Methods

apply' :: f -> a -> b

data AFst

Constructors

AFst

data ASnd

Constructors

ASnd

class HMap f (xs :: [k]) (ys :: [k]) where

Methods

hmap :: f -> HList xs -> HList ys

class HMap' f (xs :: [k]) (ys :: [k]) | f xs -> ys where

Alternative version of HMap with better type inference based on Apply`

Methods

hmap' :: f -> HList xs -> HList ys

class HMapM m f (xs :: [k]) (ys :: [k]) where

Methods

hmapM :: f -> HList xs -> m (HList ys)

class HMapM' m f (xs :: [k]) (ys :: [k]) | f xs -> ys where

Methods

hmapM' :: f -> HList xs -> m (HList ys)

class Applicative f => HSequence f (xs :: [k]) (ys :: [k]) | xs -> ys, ys f -> xs where

Methods

hsequence :: HList xs -> f (HList ys)

class HFoldr f acc xs res | f acc xs -> res where

Methods

hfoldr :: f -> acc -> HList xs -> res

class HFoldrM m f acc xs res | m f acc xs -> res where

Methods

hfoldrM :: f -> acc -> HList xs -> m res

data HNothing

Constructors

HNothing

newtype HJust x

Constructors

HJust x

class HUnfold f res xs where

Methods

hunfoldr' :: f -> res -> HList xs

type family HUnfoldRes s xs where ...

Equations

HUnfoldRes _ '[] = HNothing

HUnfoldRes s (x ': _) = HJust (x, s)

hunfoldr :: forall f res (xs :: [Type]) a. (Apply f a res, HUnfold f res xs, res ~ HUnfoldRes a xs) => f -> a -> HList xs

class HUnfoldM m f res xs where

Methods

hunfoldrM' :: f -> res -> m (HList xs)

type family HUnfoldMRes m s xs where ...

Equations

HUnfoldMRes m _ '[] = m HNothing

HUnfoldMRes m s (x ': _) = m (HJust (x, s))

hunfoldrM :: forall (m :: Type -> Type) f res (xs :: [Type]) a. (HUnfoldM m f res xs, Apply f a res, res ~ HUnfoldMRes m a xs) => f -> a -> m (HList xs)

type HReplicate n e = HReplicateFD n e (HReplicateR n e)

hreplicate :: forall n e. HReplicate n e => e -> HList (HReplicateR n e)

class HReplicateFD (n :: Nat) (e :: Type) (es :: [Type]) | n e -> es where

Methods

hreplicateFD :: e -> HList es

type family HReplicateR (n :: Nat) (e :: a) :: [a] where ...

Equations

HReplicateR 0 _ = '[]

HReplicateR n e = e ': HReplicateR (n - 1) e

type HConcat xs = HConcatFD xs (HConcatR xs)

hconcat :: HConcat xs => HList xs -> HList (HConcatR xs)

type family HConcatR (a :: [Type]) :: [Type]

type family UnHList a :: [Type]

class HConcatFD (xxs :: [k]) (xs :: [k]) | xxs -> xs where

Methods

hconcatFD :: HList xxs -> HList xs

type HAppend as bs = HAppendFD as bs (as ++ bs)

happend :: HAppend as bs => HList as -> HList bs -> HList (as ++ bs)

hunappend :: (cs ~ (as ++ bs), HAppend as bs) => HList cs -> (HList as, HList bs)

class HAppendFD (a :: [k]) (b :: [k]) (ab :: [k]) | a b -> ab, a ab -> b where

Methods

happendFD :: HList a -> HList b -> HList ab

hunappendFD :: HList ab -> (HList a, HList b)

type family (as :: [k]) ++ (bs :: [k]) :: [k] where ...

Equations

'[] ++ bs = bs

(a ': as) ++ bs = a ': (as ++ bs)

class HZip (xs :: [k]) (ys :: [k]) (zs :: [k]) | xs ys -> zs, zs -> xs ys where

Methods

hzip :: HList xs -> HList ys -> HList zs

hunzip :: HList zs -> (HList xs, HList ys)

class HZip' (xs :: [k]) (ys :: [k]) (zs :: [k]) | xs ys -> zs where

Methods

hzip' :: HList xs -> HList ys -> HList zs

data HZipF

Constructors

HZipF

htranspose :: forall (acc :: [Type]) (xs :: [Type]) (xxs :: [Type]) (res :: Type). (HReplicateFD (ListLength xs) (HList ('[] :: [Type])) acc, HFoldr HZipF (HList acc) (HList xs ': xxs) res) => HList (HList xs ': xxs) -> res

class HZipWith f (xs :: [k]) (ys :: [k]) (zs :: [k]) | f xs ys -> zs where

Methods

hzipWith :: f -> HList xs -> HList ys -> HList zs

class HZipWithM m f (xs :: [k]) (ys :: [k]) (zs :: [k]) | f xs ys -> zs where

Methods

hzipWithM :: f -> HList xs -> HList ys -> m (HList zs)

class HZip3 (as :: [k]) (bs :: [k]) (cs :: [k]) (ds :: [k]) | as bs cs -> ds, ds -> as bs cs where

Methods

hzip3 :: HList as -> HList bs -> HList cs -> HList ds

hunzip3 :: HList ds -> (HList as, HList bs, HList cs)

class HZipWith3 f (as :: [k]) (bs :: [k]) (cs :: [k]) (ds :: [k]) | f as bs cs -> ds where

Methods

hzipWith3 :: f -> HList as -> HList bs -> HList cs -> HList ds

class HCartesianProduct (xs :: [k]) (ys :: [k]) (zs :: [k]) | xs ys -> zs where

Methods

hproduct :: HList xs -> HList ys -> HList zs

class HAttach x (ys :: [k]) (zs :: [k]) | x ys -> zs where

Methods

hattach :: x -> HList ys -> HList zs