Torch.Optim

Documentation

type LearningRate = Tensor

type Loss = Tensor

newtype Gradients

Constructors

Gradients [Tensor]

newtype OptimizerState option

Constructors

OptimizerState option

grad' :: Loss -> [Parameter] -> Gradients

class Optimizer optimizer where

Minimal complete definition

step

Methods

step :: LearningRate -> Gradients -> [Tensor] -> optimizer -> ([Tensor], optimizer)

runStep :: Parameterized model => model -> optimizer -> Loss -> LearningRate -> IO (model, optimizer)

run a single iteration of an optimizer, returning new parameters and updated optimizer state

runStep' :: Parameterized model => model -> optimizer -> Gradients -> LearningRate -> IO (model, optimizer)

run a single iteration of an optimizer, returning new parameters and updated optimizer state

data GD

Constructors

GD

gd :: LearningRate -> Gradients -> [Tensor] -> [Tensor]

Stateless gradient descent step

gd' :: LearningRate -> Gradients -> [Tensor] -> GD -> ([Tensor], GD)

Gradient descent step with a dummy state variable

sgd :: LearningRate -> [Parameter] -> [Tensor] -> [Tensor]

data GDM

Constructors

GDM

Fields

beta :: Float

momentum :: [Tensor]

gdm

Arguments

:: LearningRate
learning rate

-> Gradients
model parameter gradients

-> [Tensor]
model parameters

-> GDM
beta & momentum

-> ([Tensor], GDM)
returns new parameters + updated momentum

data Adam

State representation for Adam Optimizer

Constructors

Adam

Fields

beta1 :: Float

beta2 :: Float

m1 :: [Tensor]

m2 :: [Tensor]

iter :: Int

mkAdam :: Int -> Float -> Float -> [Parameter] -> Adam

adam

Arguments

:: LearningRate
learning rate

-> Gradients
model parameter gradients

-> [Tensor]
model parameters

-> Adam
adam parameters - beta1, beta2, moments, iteration

-> ([Tensor], Adam)
returns new parameters + updated adam parameters

Adam step

foldLoop :: a -> Int -> (a -> Int -> IO a) -> IO a

syntactic sugar for looping with foldM