Torch.Tensor

Documentation

type ATenTensor = ForeignPtr Tensor

newtype Tensor

Constructors

Unsafe ATenTensor

numel

Arguments

:: Tensor
input

-> Int
number of elements in tensor

Returns the total number of elements in the input tensor.

size :: Tensor -> Int -> Int

shape

Arguments

:: Tensor
input

-> [Int]
list of integers representing the shape of the tensor

Returns the shape of the tensor

dim

Arguments

:: Tensor
input

-> Int
output

Returns the dimensions of the input tensor

device

Arguments

:: Tensor
input

-> Device
object representing the device

Returns the device on which the tensor is currently allocated

dtype

Arguments

:: Tensor
input

-> DType
data type of the input tensor

Returns the data type of the input tensor

toDouble :: Tensor -> Double

toInt :: Tensor -> Int

toType

Arguments

:: DType
data type to cast input to

-> Tensor
input

-> Tensor
output

Casts the input tensor to the given data type

toDevice

Arguments

:: Device
device to cast input to

-> Tensor
input

-> Tensor
output

Casts the input tensor to given device

select

Arguments

:: Tensor
input

-> Int
dimension to slice along

-> Int
index in the given dimension

-> Tensor
output

Slices the input tensor along the selected dimension at the given index.

indexSelect :: Tensor -> Int -> Tensor -> Tensor

Returns a new tensor which indexes the input tensor along dimension dim using the entries in index which is a LongTensor.

reshape :: [Int] -> Tensor -> Tensor

Returns a tensor with the same data and number of elements as input, but with the specified shape.

toSparse :: Tensor -> Tensor

toDense :: Tensor -> Tensor

toMKLDNN :: Tensor -> Tensor

toCPU :: Tensor -> Tensor

toCUDA :: Tensor -> Tensor

(@@) :: TensorIndex a => Tensor -> a -> Tensor

class TensorIndex a where

Methods

indexWith :: (Tensor, Int) -> a -> (Tensor, Int)

class TensorLike a where

Methods

asTensor' :: a -> TensorOptions -> Tensor

asTensor :: a -> Tensor

asValue :: Tensor -> a

_dtype :: DType

_dims :: a -> [Int]

_deepDims :: a -> Maybe [Int]

_peekElemOff :: Ptr () -> Int -> [Int] -> IO a

_pokeElemOff :: Ptr () -> Int -> a -> IO ()

int64_opts :: TensorOptions

float_opts :: TensorOptions

withTensor :: Tensor -> (Ptr () -> IO a) -> IO a