Torch.TensorFactories

Documentation

type FactoryType = ForeignPtr IntArray -> ForeignPtr TensorOptions -> IO (ForeignPtr Tensor)

type FactoryTypeWithDimnames = ForeignPtr IntArray -> ForeignPtr DimnameList -> ForeignPtr TensorOptions -> IO (ForeignPtr Tensor)

mkFactory

Arguments

:: FactoryType
aten_impl

-> [Int]
shape

-> TensorOptions
opts

-> IO Tensor
output

mkFactoryUnsafe :: FactoryType -> [Int] -> TensorOptions -> Tensor

mkFactoryWithDimnames :: FactoryTypeWithDimnames -> [(Int, Dimname)] -> TensorOptions -> IO Tensor

mkFactoryUnsafeWithDimnames :: FactoryTypeWithDimnames -> [(Int, Dimname)] -> TensorOptions -> Tensor

mkDefaultFactory :: ([Int] -> TensorOptions -> a) -> [Int] -> a

mkDefaultFactoryWithDimnames :: ([(Int, Dimname)] -> TensorOptions -> a) -> [(Int, Dimname)] -> a

ones

Arguments

:: [Int]
sequence of integers defining the shape of the output tensor.

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> Tensor
output

Returns a tensor filled with the scalar value 1, with the shape defined by the variable argument size.

onesLike

Arguments

:: Tensor
input

-> Tensor
output

Returns a tensor filled with the scalar value 1, with the same size as input tensor

zeros

Arguments

:: [Int]
sequence of integers defining the shape of the output tensor.

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> Tensor
output

Returns a tensor filled with the scalar value 0, with the shape defined by the variable argument size.

zerosLike

Arguments

:: Tensor
input

-> Tensor
output

Returns a tensor filled with the scalar value 0, with the same size as input tensor

randIO

Arguments

:: [Int]
sequence of integers defining the shape of the output tensor.

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> IO Tensor
output

Returns a tensor filled with random numbers from a uniform distribution on the interval [0,1)

randnIO

Arguments

:: [Int]
sequence of integers defining the shape of the output tensor.

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> IO Tensor
output

Returns a tensor filled with random numbers from a standard normal distribution.

randintIO

Arguments

:: Int
lowest integer to be drawn from the distribution. Default: 0.

-> Int
one above the highest integer to be drawn from the distribution.

-> [Int]
the shape of the output tensor.

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> IO Tensor
output

Returns a tensor filled with random integers generated uniformly between low (inclusive) and high (exclusive).

randnLikeIO

Arguments

:: Tensor
input

-> IO Tensor
output

Returns a tensor with the same size as input that is filled with random numbers from standard normal distribution.

randLikeIO

Arguments

:: Tensor
input

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> IO Tensor
output

Returns a tensor with the same size as input that is filled with random numbers from a uniform distribution on the interval [0,1).

fullLike

Arguments

:: Tensor
input

-> Float
_fill_value

-> TensorOptions
opt

-> IO Tensor
output

onesWithDimnames :: [(Int, Dimname)] -> TensorOptions -> Tensor

zerosWithDimnames :: [(Int, Dimname)] -> TensorOptions -> Tensor

randWithDimnames :: [(Int, Dimname)] -> TensorOptions -> IO Tensor

randnWithDimnames :: [(Int, Dimname)] -> TensorOptions -> IO Tensor

linspace

Arguments

:: (Scalar a, Scalar b)

=> a
start

-> b
end

-> Int
steps

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> Tensor
output

Returns a one-dimensional tensor of steps equally spaced points between start and end.

logspace :: (Scalar a, Scalar b) => a -> b -> Int -> Double -> TensorOptions -> Tensor

eyeSquare

Arguments

:: Int
dim

-> TensorOptions
opts

-> Tensor
output

eye

Arguments

:: Int
the number of rows

-> Int
the number of columns

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> Tensor
output

Returns a 2-D tensor with ones on the diagonal and zeros elsewhere.

full

Arguments

:: Scalar a

=> [Int]
the shape of the output tensor.

-> a
the number to fill the output tensor with

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> Tensor
output

Returns a tensor of given size filled with fill_value.

sparseCooTensor

Arguments

:: Tensor
The indices are the coordinates of the non-zero values in the matrix

-> Tensor
Initial values for the tensor.

-> [Int]
the shape of the output tensor.

-> TensorOptions

-> Tensor
output

Constructs a sparse tensors in COO(rdinate) format with non-zero elements at the given indices with the given values.

ones' :: [Int] -> Tensor

zeros' :: [Int] -> Tensor

randIO' :: [Int] -> IO Tensor

randnIO' :: [Int] -> IO Tensor

randintIO' :: Int -> Int -> [Int] -> IO Tensor

randLikeIO' :: Tensor -> IO Tensor

bernoulliIO'

Arguments

:: Tensor
t

-> IO Tensor
output

bernoulliIO

Arguments

:: Tensor
t

-> Double
p

-> IO Tensor
output

poissonIO

Arguments

:: Tensor
t

-> IO Tensor
output

multinomialIO'

Arguments

:: Tensor
t

-> Int
num_samples

-> IO Tensor
output

multinomialIO

Arguments

:: Tensor
t

-> Int
num_samples

-> Bool
replacement

-> IO Tensor
output

normalIO'

Arguments

:: Tensor
_mean

-> IO Tensor
output

normalIO

Arguments

:: Tensor
_mean

-> Tensor
_std

-> IO Tensor
output

normalScalarIO

Arguments

:: Tensor
_mean

-> Double
_std

-> IO Tensor
output

normalScalarIO'

Arguments

:: Double
_mean

-> Tensor
_std

-> IO Tensor
output

normalWithSizeIO

Arguments

:: Double
_mean

-> Double
_std

-> Int
_size

-> IO Tensor
output

rreluIO'''

Arguments

:: Tensor
t

-> IO Tensor
output

rreluIO''

Arguments

:: Scalar a

=> Tensor
t

-> a
upper

-> IO Tensor
output

rreluIO'

Arguments

:: Scalar a

=> Tensor
t

-> a
lower

-> a
upper

-> IO Tensor
output

rreluIO

Arguments

:: Scalar a

=> Tensor
t

-> a
lower

-> a
upper

-> Bool
training

-> IO Tensor
output

rreluWithNoiseIO'''

Arguments

:: Tensor
t

-> Tensor
noise

-> IO Tensor
output

rreluWithNoiseIO''

Arguments

:: Scalar a

=> Tensor
t

-> Tensor
noise

-> a
upper

-> IO Tensor
output

rreluWithNoiseIO'

Arguments

:: Scalar a

=> Tensor
t

-> Tensor
noise

-> a
lower

-> a
upper

-> IO Tensor
output

rreluWithNoiseIO

Arguments

:: Scalar a

=> Tensor
t

-> Tensor
noise

-> a
lower

-> a
upper

-> Bool
training

-> IO Tensor
output

onesWithDimnames' :: [(Int, Dimname)] -> Tensor

zerosWithDimnames' :: [(Int, Dimname)] -> Tensor

randWithDimnames' :: [(Int, Dimname)] -> IO Tensor

randnWithDimnames' :: [(Int, Dimname)] -> IO Tensor

linspace' :: (Scalar a, Scalar b) => a -> b -> Int -> Tensor

logspace' :: (Scalar a, Scalar b) => a -> b -> Int -> Double -> Tensor

eyeSquare' :: Int -> Tensor

eye' :: Int -> Int -> Tensor

full' :: Scalar a => [Int] -> a -> Tensor

sparseCooTensor' :: Tensor -> Tensor -> [Int] -> Tensor

arange

Arguments

:: Int
start

-> Int
end

-> Int
step

-> TensorOptions
configures the data type, device, layout and other properties of the resulting tensor.

-> Tensor
output

Returns a 1-D tensor with values from the interval [start, end) taken with common difference step beginning from start.

arange'

Arguments

:: Int
start

-> Int
end

-> Int
step

-> Tensor
output

Returns a 1-D tensor with values from the interval [start, end) taken with common difference step beginning from start.