Torch.Typed.NN

Documentation

class HasForward f a b | f a -> b where

Minimal complete definition

forward

Methods

forward :: f -> a -> b

forwardStoch :: f -> a -> IO b

data LinearSpec (inputFeatures :: Nat) (outputFeatures :: Nat) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

LinearSpec

data Linear (inputFeatures :: Nat) (outputFeatures :: Nat) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

Linear

Fields

:: forall inputFeatures outputFeatures dtype device. { linearWeight :: Parameter device dtype '[outputFeatures, inputFeatures]

, linearBias :: Parameter device dtype '[outputFeatures]

} -> Linear inputFeatures outputFeatures dtype device

linear :: _ => Linear _ _ _ _ -> Tensor _ _ _ -> Tensor _ _ _

linear The constraints on this one are _very_ involved, so the partial signatures make the code significantly cleaner.

data DropoutSpec where

Constructors

DropoutSpec

Fields

:: { dropoutProbSpec :: Double

} -> DropoutSpec

data Dropout where

Constructors

Dropout

Fields

:: { dropoutProb :: Double

} -> Dropout

dropout :: forall shape dtype device. Dropout -> Bool -> Tensor device dtype shape -> IO (Tensor device dtype shape)

data EmbeddingType

Constructors

Constant

Learned

data EmbeddingSpec (paddingIdx :: Maybe Nat) (numEmbeds :: Nat) (embedSize :: Nat) (embeddingType :: EmbeddingType) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

ConstEmbeddingSpec :: forall paddingIdx numEmbeds embedSize dtype device. Tensor device dtype '[numEmbeds, embedSize] -> EmbeddingSpec paddingIdx numEmbeds embedSize 'Constant dtype device

LearnedEmbeddingWithRandomInitSpec :: forall paddingIdx numEmbeds embedSize dtype device. EmbeddingSpec paddingIdx numEmbeds embedSize 'Learned dtype device

LearnedEmbeddingWithCustomInitSpec :: forall paddingIdx numEmbeds embedSize dtype device. Tensor device dtype '[numEmbeds, embedSize] -> EmbeddingSpec paddingIdx numEmbeds embedSize 'Learned dtype device

data Embedding (paddingIdx :: Maybe Nat) (numEmbeds :: Nat) (embedSize :: Nat) (embeddingType :: EmbeddingType) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

ConstEmbedding

Fields

:: forall paddingIdx numEmbeds embedSize dtype device. { constEmbedWeights :: Tensor device dtype '[numEmbeds, embedSize]

} -> Embedding paddingIdx numEmbeds embedSize 'Constant dtype device

LearnedEmbedding

Fields

:: forall paddingIdx numEmbeds embedSize dtype device. { learnedEmbedWeights :: Parameter device dtype '[numEmbeds, embedSize]

} -> Embedding paddingIdx numEmbeds embedSize 'Learned dtype device

embed :: forall paddingIdx shape numEmbeds embedSize embeddingType dtype device shape'. (KnownMaybeNat paddingIdx, PaddingIdxCheck paddingIdx numEmbeds, shape' ~ Reverse (embedSize ': Reverse shape)) => Embedding paddingIdx numEmbeds embedSize embeddingType dtype device -> Tensor device 'Int64 shape -> Tensor device dtype shape'

data Conv1dSpec (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize :: Nat) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

Conv1dSpec

data Conv1d (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize :: Nat) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

Conv1d

Fields

:: forall inputChannelSize outputChannelSize kernelSize dtype device. { conv1dWeight :: Parameter device dtype '[outputChannelSize, inputChannelSize, kernelSize]

, conv1dBias :: Parameter device dtype '[outputChannelSize]

} -> Conv1d inputChannelSize outputChannelSize kernelSize dtype device

conv1d :: forall stride padding. _ => Conv1d _ _ _ _ _ -> Tensor _ _ _ -> Tensor _ _ _

conv1d The constraints on this one are _very_ involved, so the partial signatures make the code significantly cleaner.

data Conv2dSpec (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize0 :: Nat) (kernelSize1 :: Nat) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

Conv2dSpec

data Conv2d (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize0 :: Nat) (kernelSize1 :: Nat) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

Conv2d

Fields

:: forall inputChannelSize outputChannelSize kernelSize0 kernelSize1 dtype device. { conv2dWeight :: Parameter device dtype '[outputChannelSize, inputChannelSize, kernelSize0, kernelSize1]

, conv2dBias :: Parameter device dtype '[outputChannelSize]

} -> Conv2d inputChannelSize outputChannelSize kernelSize0 kernelSize1 dtype device

conv2d :: forall stride padding. _ => Conv2d _ _ _ _ _ _ -> Tensor _ _ _ -> Tensor _ _ _

conv2d The constraints on this one are _very_ involved, so the partial signatures make the code significantly cleaner.

data Conv3dSpec (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize0 :: Nat) (kernelSize1 :: Nat) (kernelSize2 :: Nat) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

Conv3dSpec

data Conv3d (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize0 :: Nat) (kernelSize1 :: Nat) (kernelSize2 :: Nat) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

Conv3d

Fields

:: forall inputChannelSize outputChannelSize kernelSize0 kernelSize1 kernelSize2 dtype device. { conv3dWeight :: Parameter device dtype '[outputChannelSize, inputChannelSize, kernelSize0, kernelSize1, kernelSize2]

, conv3dBias :: Parameter device dtype '[outputChannelSize]

} -> Conv3d inputChannelSize outputChannelSize kernelSize0 kernelSize1 kernelSize2 dtype device

conv3d :: forall stride padding. _ => Conv3d _ _ _ _ _ _ _ -> Tensor _ _ _ -> Tensor _ _ _

conv3d The constraints on this one are _very_ involved, so the partial signatures make the code significantly cleaner.

data LayerNormSpec (normalizedShape :: [Nat]) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

LayerNormSpec

Fields

:: forall normalizedShape dtype device. { layerNormEpsSpec :: Double

} -> LayerNormSpec normalizedShape dtype device

data LayerNorm (normalizedShape :: [Nat]) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

LayerNorm

Fields

:: { layerNormWeight :: Parameter device dtype normalizedShape

, layerNormBias :: Parameter device dtype normalizedShape

, layerNormEps :: Double

} -> LayerNorm normalizedShape dtype device

layerNorm :: forall normalizedShape shape dtype device. (EndsWith shape normalizedShape, KnownShape normalizedShape) => LayerNorm normalizedShape dtype device -> Tensor device dtype shape -> Tensor device dtype shape