Torch.Typed.NN.Convolution

Documentation

data Conv1dSpec (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize :: Nat) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

Conv1dSpec

data Conv1d (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize :: Nat) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

Conv1d

Fields

:: forall inputChannelSize outputChannelSize kernelSize dtype device. { conv1dWeight :: Parameter device dtype '[outputChannelSize, inputChannelSize, kernelSize]

, conv1dBias :: Parameter device dtype '[outputChannelSize]

} -> Conv1d inputChannelSize outputChannelSize kernelSize dtype device

conv1dForward :: forall stride padding. _ => Conv1d _ _ _ _ _ -> Tensor _ _ _ -> Tensor _ _ _

conv1d The constraints on this one are _very_ involved, so the partial signatures make the code significantly cleaner.

data Conv2dSpec (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize0 :: Nat) (kernelSize1 :: Nat) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

Conv2dSpec

data Conv2d (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize0 :: Nat) (kernelSize1 :: Nat) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

Conv2d

Fields

:: forall inputChannelSize outputChannelSize kernelSize0 kernelSize1 dtype device. { conv2dWeight :: Parameter device dtype '[outputChannelSize, inputChannelSize, kernelSize0, kernelSize1]

, conv2dBias :: Parameter device dtype '[outputChannelSize]

} -> Conv2d inputChannelSize outputChannelSize kernelSize0 kernelSize1 dtype device

conv2dForward :: forall stride padding. _ => Conv2d _ _ _ _ _ _ -> Tensor _ _ _ -> Tensor _ _ _

conv2d The constraints on this one are _very_ involved, so the partial signatures make the code significantly cleaner.

data Conv3dSpec (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize0 :: Nat) (kernelSize1 :: Nat) (kernelSize2 :: Nat) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

Conv3dSpec

data Conv3d (inputChannelSize :: Nat) (outputChannelSize :: Nat) (kernelSize0 :: Nat) (kernelSize1 :: Nat) (kernelSize2 :: Nat) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

Conv3d

Fields

:: forall inputChannelSize outputChannelSize kernelSize0 kernelSize1 kernelSize2 dtype device. { conv3dWeight :: Parameter device dtype '[outputChannelSize, inputChannelSize, kernelSize0, kernelSize1, kernelSize2]

, conv3dBias :: Parameter device dtype '[outputChannelSize]

} -> Conv3d inputChannelSize outputChannelSize kernelSize0 kernelSize1 kernelSize2 dtype device

conv3dForward :: forall stride padding. _ => Conv3d _ _ _ _ _ _ _ -> Tensor _ _ _ -> Tensor _ _ _

conv3d The constraints on this one are _very_ involved, so the partial signatures make the code significantly cleaner.