Torch.Functional

Documentation

data KeepDim

Constructors

KeepDim

RemoveDim

data Dim

Constructors

Dim Int

data Reduction

Constructors

ReduceNone

ReduceMean

ReduceSum

data Tri

Constructors

Upper

Lower

kOne :: ForeignPtr Scalar

isUpper :: Tri -> Bool

mean

Arguments

:: Tensor
input

-> Tensor
output

Returns the mean value of all elements in the input tensor.

std

Arguments

:: Tensor
input

-> Tensor
output

Returns the standard deviation of all elements in the input tensor.

var

Arguments

:: Tensor
input

-> Tensor
output

Returns the variance of all elements in the input tensor.

sumAll

Arguments

:: Tensor
input

-> Tensor
output

Returns the sum of all elements in the input tensor.

abs

Arguments

:: Tensor
input

-> Tensor
output

Computes the element-wise absolute value of the given input tensor.

frac

Arguments

:: Tensor
input

-> Tensor
output

Computes the fractional portion of each element in input. out_i = input_i - (floor . abs) input_i * (sign input_i)

keepdim :: KeepDim -> Bool

argmax

Arguments

:: Dim
the dimension to reduce

-> KeepDim
whether the output tensor has dim retained or not

-> Tensor
input

-> Tensor
output

Returns the indices of the maximum value of all elements in the input tensor.

add

Arguments

:: Tensor
input

-> Tensor
other

-> Tensor
output

Each element of the tensor other added to each element of the tensor input. The resulting tensor is returned.

mul

Arguments

:: Tensor
input

-> Tensor
other

-> Tensor
output

Multiplies each element of the tensor other to each element of the input tensor and returns a new resulting tensor.

sub

Arguments

:: Tensor
input

-> Tensor
other

-> Tensor
output

Element wise subtraction of other tensor from input tensor and returns a new resulting tensor

ceil

Arguments

:: Tensor
input

-> Tensor
output

ceil

floor

Arguments

:: Tensor
input

-> Tensor
output

floor

min

Arguments

:: Tensor
input

-> Tensor
output

min

max

Arguments

:: Tensor
input

-> Tensor
output

max

median

Arguments

:: Tensor
input

-> Tensor
output

median

addScalar

Arguments

:: Scalar a

=> Tensor
input

-> a
summand

-> Tensor
output

Adds each element of the input input with the scalar and returns a new resulting tensor.

subScalar

Arguments

:: Scalar a

=> Tensor
input

-> a
subtrahend

-> Tensor
output

Subtracts each element of the input input with the scalar and returns a new resulting tensor.

mulScalar

Arguments

:: Scalar a

=> Tensor
input

-> a
multiplier

-> Tensor
output

Multiplies each element of the input input with the scalar and returns a new resulting tensor.

divScalar

Arguments

:: Scalar a

=> Tensor
input

-> a
divisor

-> Tensor
output

Divides each element of the input input with the scalar and returns a new resulting tensor.

matmul

Arguments

:: Tensor
first tensor for matrix multiplication

-> Tensor
second tensor for matrix multiplication

-> Tensor
output

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned. If both arguments are 2-dimensional, the matrix-matrix product is returned. If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed. If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned. If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a (j times 1 times n times m)(j×1×n×m) tensor and other is a (k times m times p)(k×m×p) tensor, out will be an (j times k times n times p)(j×k×n×p) tensor.

erf

Arguments

:: Tensor
input

-> Tensor
output

Computes the error function of each element

exp

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the exponential of the elements of the input tensor input.

log1p :: Tensor -> Tensor

Returns a new tensor with the natural logarithm of (1 + input).

log2

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the logarithm to the base 2 of the elements of input.

log10

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the logarithm to the base 10 of the elements of input.

pow

Arguments

:: Scalar a

=> Tensor
input

-> a
exponent

-> Tensor
output

Takes the power of each element in input with exponent and returns a tensor with the result.

powt

Arguments

:: Tensor
input

-> Tensor
exponent

-> Tensor
output

Takes the power of each element in input with exponent and returns a tensor with the result. Exponent is a tensor with the same number of elements as input.

relu

Arguments

:: Tensor
input

-> Tensor
output

Applies the rectified linear unit function element-wise.

selu

Arguments

:: Tensor
input

-> Tensor
output

Applies element-wise, \(\text{SELU}(x) = scale * (\max(0,x) + \min(0, \alpha * (\exp(x) - 1))\) , with α=1.6732632423543772848170429916717 and scale=1.0507009873554804934193349852946.

celu

Arguments

:: Float
alpha

-> Tensor
input

-> Tensor
output

Applies element-wise, \(\text{CELU}(x) = \max(0,x) + \min(0, \alpha * (\exp(x/\alpha) - 1))\).

sigmoid

Arguments

:: Tensor
input

-> Tensor
output

Applies the element-wise function sigmoid.

softmax

Arguments

:: Int
dimension

-> Tensor
input

-> Tensor
output

Applies a softmax function. It is applied to all slices along dim, and will re-scale them so that the elements lie in the range [0, 1] and sum to 1.

logSoftmax

Arguments

:: Int
dimension

-> Tensor
input

-> Tensor
output

Applies a softmax followed by a logarithm. While mathematically equivalent to log(softmax(x)), doing these two operations separately is slower, and numerically unstable. This function uses an alternative formulation to compute the output and gradient correctly.

threshold

Arguments

:: Float
threshold

-> Float
value

-> Tensor
input

-> Tensor
output

Thresholds each element of the input Tensor.

sin

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the sine of the elements of input.

cos

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the cos of the elements of input.

tan

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the tangent of the elements of input.

sinh

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the hyperbolic sine of the elements of input.

cosh

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the hyperbolic cosine of the elements of input.

tanh

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the hyperbolic tangent of the elements of input.

sqrt

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the square-root of the elements of input.

gt

Arguments

:: Tensor
input

-> Tensor
output

-> Tensor
other

Computes input > other element-wise. The second argument can be a number or a tensor whose shape is broadcastable with the first argument.

(>.) :: Tensor -> Tensor -> Tensor

lt

Arguments

:: Tensor
input

-> Tensor
other

-> Tensor
output

Computes input < other element-wise. The second argument can be a number or a tensor whose shape is broadcastable with the first argument.

(<.) :: Tensor -> Tensor -> Tensor

ge

Arguments

:: Tensor
input

-> Tensor
other

-> Tensor
output

Computes input >= other element-wise. The second argument can be a number or a tensor whose shape is broadcastable with the first argument.

(>=.) :: Tensor -> Tensor -> Tensor

le

Arguments

:: Tensor
input

-> Tensor
other

-> Tensor
output

Computes input <= other element-wise. The second argument can be a number or a tensor whose shape is broadcastable with the first argument.

(<=.) :: Tensor -> Tensor -> Tensor

eq

Arguments

:: Tensor
input

-> Tensor
other

-> Tensor
output

Computes input == other element-wise. The second argument can be a number or a tensor whose shape is broadcastable with the first argument.

(==.) :: Tensor -> Tensor -> Tensor

ne

Arguments

:: Tensor
input

-> Tensor
other

-> Tensor
output

Computes input /= other element-wise. The second argument can be a number or a tensor whose shape is broadcastable with the first argument.

(/=.) :: Tensor -> Tensor -> Tensor

toDType

Arguments

:: DType
data type to cast to

-> Tensor
input

-> Tensor
output

Casting to given Dtype, where Dtype is an object that represents the data type of a tensor in hasktorch.

squeezeAll

Arguments

:: Tensor
input

-> Tensor
output

squeezeAll

binaryCrossEntropyLoss

Arguments

:: Tensor
input

-> Tensor
target

-> Tensor
weight

-> Reduction
Specifies the reduction to apply to the output

-> Tensor
output

Function that measures the Binary Cross Entropy between the target and the output.

binaryCrossEntropyLoss'

Arguments

:: Tensor
input

-> Tensor
target

-> Tensor
output

Binary Cross Entropy with weights defaulted to 1.0 & reduction defaulted to ReduceMean

mseLoss

Arguments

:: Tensor
input

-> Tensor
target

-> Tensor
output

Creates a criterion that measures the mean squared error (squared L2 norm) between each element in the input and target.

nllLoss'

Arguments

:: Tensor
input

-> Tensor
target tensor

-> Tensor
output

The negative log likelihood loss.

adaptiveMaxPool1d

Arguments

:: Int
output size

-> Tensor
input

-> (Tensor, Tensor)
output

Applies a 1D adaptive max pooling over an input signal composed of several input planes.

adaptiveMaxPool2d

Arguments

:: (Int, Int)
output size

-> Tensor
input

-> (Tensor, Tensor)
output

Applies a 2D adaptive max pooling over an input signal composed of several input planes.

maxPool1dWithIndices

Arguments

:: Int
kernel size

-> Int
stride

-> Int
padding

-> Int
dilation

-> Bool
ceil mode

-> Tensor
input

-> (Tensor, Tensor)
output, indices

maxPool1dWithIndices

maxPool1d

Arguments

:: Int
kernel size

-> Int
stride

-> Int
padding

-> Int
dilation

-> Bool
ceil mode

-> Tensor
input

-> Tensor
output

Applies a 1D max pooling over an input signal composed of several input planes.

maxPool2d

Arguments

:: (Int, Int)
kernel size

-> (Int, Int)
stride

-> (Int, Int)
padding

-> (Int, Int)
dilation

-> Bool
ceil mode

-> Tensor
input

-> Tensor
output

Applies a 2D max pooling over an input signal composed of several input planes.

maxPool3d

Arguments

:: (Int, Int, Int)
kernel size

-> (Int, Int, Int)
stride

-> (Int, Int, Int)
padding

-> (Int, Int, Int)
dilation

-> Bool
ceil mode

-> Tensor
input

-> Tensor
output

Applies a 3D max pooling over an input signal composed of several input planes.

inverse

Arguments

:: Tensor
input

-> Tensor
output

Takes the inverse of the square matrix input. input can be batches of 2D square tensors, in which case this function would return a tensor composed of individual inverses.

symeig

Arguments

:: Tensor
input tensor

-> Bool
bool which controls whether eigenvectors have to be computed

-> Tri
controls whether to consider upper-triangular or lower-triangular region

-> (Tensor, Tensor)
output tensors

This function returns eigenvalues and eigenvectors of a real symmetric matrix input or a batch of real symmetric matrices, represented by a namedtuple (eigenvalues, eigenvectors).

eig

Arguments

:: Tensor
input (square matrix) for which the eigen values and eigen vectors are to be computed

-> Bool
bool to compute both eigenvalues and eigenvectors; otherwise, only eigenvalues will be computed

-> (Tensor, Tensor)
output tensors

Computes the eigenvalues and eigenvectors of a real square matrix

svd

Arguments

:: Tensor
input

-> Bool
controls the shape of returned U and V

-> Bool
option whether to compute U and V or not

-> (Tensor, Tensor, Tensor)
output tuple of tensors

This function returns a namedtuple (U, S, V) which is the singular value decomposition of a input real matrix or batches of real matrices input such that input = U * diag(S) * V^T

cholesky

Arguments

:: Tensor
input

-> Tri
flag that indicates whether to return a upper or lower triangular matrix.

-> Tensor
output

Computes the Cholesky decomposition of a symmetric positive-definite matrix AA or for batches of symmetric positive-definite matrices.

choleskySolve

Arguments

:: Tensor
input matrix b

-> Tensor
input matrix u

-> Tri
bool whether to consider the Cholesky factor as a lower or upper triangular matrix

-> Tensor
output

Solves a linear system of equations with a positive semidefinite matrix to be inverted given its Cholesky factor matrix uu .

dropout

Arguments

:: Double
dropout probability

-> Bool
whether or not to activate dropout

-> Tensor
input

-> IO Tensor
output

During training, randomly zeroes some of the elements of the input tensor with probability p using samples from a Bernoulli distribution.

featureDropout

Arguments

:: Double
dropout probability

-> Bool
whether or not to activate dropout

-> Tensor
input

-> IO Tensor
output

alphaDropout

Arguments

:: Double
dropout probability

-> Bool
whether or not to activate dropout

-> Tensor
input

-> IO Tensor
output

Applies alpha dropout to the input.

featureAlphaDropout

Arguments

:: Double
dropout probability

-> Bool
whether or not to activate dropout

-> Tensor
input

-> IO Tensor
output

avgPool1d

Arguments

:: Int
kernel size

-> Int
stride

-> Int
padding

-> Bool
ceil mode

-> Bool
count include pad

-> Tensor
input

-> Tensor
output

Applies a 1D average pooling over an input signal composed of several input planes.

avgPool1d'

Arguments

:: Int
kernel size

-> Int
stride

-> Int
padding

-> Tensor
input

-> Tensor
output

adaptiveAvgPool1d

Arguments

:: Int

-> Tensor
input

-> Tensor
output

Applies a 1D adaptive average pooling over an input signal composed of several input planes.

adaptiveAvgPool2d

Arguments

:: Tensor
input

-> (Int, Int)
output size (Height * Width)

-> Tensor
output

Applies a 2D adaptive average pooling over an input signal composed of several input planes.

adaptiveAvgPool3d

Arguments

:: Tensor
input

-> (Int, Int, Int)
output size (Depth * Height * Width)

-> Tensor
output

Applies a 3D adaptive average pooling over an input signal composed of several input planes.

bitwiseNot

Arguments

:: Tensor
input

-> Tensor
output

Computes the bitwise NOT of the given input tensor. The input tensor must be of integral or Boolean types. For bool tensors, it computes the logical NOT.

cat

Arguments

:: Int
dimension

-> [Tensor]
list of tensors to concatenate

-> Tensor
output tensor

Concatenates the given sequence of seq tensors in the given dimension. All tensors must either have the same shape (except in the concatenating dimension) or be empty.

chunk

Arguments

:: Int
chunks

-> Int
dim

-> Tensor
input tensor

-> [Tensor]
output list of tensors

Splits a tensor into a specific number of chunks. Last chunk will be smaller if the tensor size along the given dimension dim is not divisible by chunks.

clamp

Arguments

:: Float
minimum value

-> Float
maximum value

-> Tensor
input

-> Tensor
output

Clamp all elements in input into the range [ min, max ] and return a resulting tensor.

clampMax

Arguments

:: Float
maximum value

-> Tensor
input

-> Tensor
output

Clamps all elements in input to be smaller or equal max.

clampMin

Arguments

:: Float
minimum value

-> Tensor
input

-> Tensor
output

Clamps all elements in input to be larger or equal min.

cudnnIsAcceptable

Arguments

:: Tensor
input

-> Bool
output

constantPadNd1d

Arguments

:: [Int]
list of padding per dimension

-> Float
value

-> Tensor
input

-> Tensor
ouptut

Pads the input tensor boundaries with a constant value.

conv1d

Arguments

:: Tensor
weight

-> Tensor
bias

-> Int
stride

-> Int
padding

-> Int
dilation

-> Int
groups

-> Tensor
input

-> Tensor
output

Applies a 1D convolution over an input signal composed of several input planes.

conv1d' :: Tensor -> Tensor -> Int -> Int -> Tensor -> Tensor

conv2d

Arguments

:: Tensor
weight

-> Tensor
bias

-> (Int, Int)
strides

-> (Int, Int)
padding

-> (Int, Int)
dilation

-> Int
groups

-> Tensor
input

-> Tensor
output

Applies a 2D convolution over an input signal composed of several input planes.

conv2d'

Arguments

:: Tensor
weight

-> Tensor
bias

-> (Int, Int)
strides

-> (Int, Int)
padding

-> Tensor
input

-> Tensor
output

solve

Arguments

:: Tensor
input matrix

-> Tensor
input square matrix

-> (Tensor, Tensor)
output tuple with solution and LU

This function returns the solution to the system of linear equations represented by AX = BAX=B and the LU factorization of A, in order as a namedtuple solution, LU. LU contains L and U factors for LU factorization of A

choleskyInverse

Arguments

:: Tensor
input

-> Tri
upper or lower triangle

-> Tensor
solution

Solves a linear system of equations with a positive semidefinite matrix to be inverted given its Cholesky factor matrix uu .

geqrf

Arguments

:: Tensor
input

-> (Tensor, Tensor)
a, tau output matrices (see https://software.intel.com/en-us/node/521004)

This is a low-level function for calling LAPACK directly. This function returns a namedtuple (a, tau) as defined in LAPACK documentation for geqrf.

orgqr

Arguments

:: Tensor
the a from geqrf function

-> Tensor
the tau from geqrf function

-> Tensor
output

Computes the orthogonal matrix Q of a QR factorization, from the (input, input2) tuple returned by geqrf function. This directly calls the underlying LAPACK function ?orgqr. See LAPACK documentation for orgqr for further details.

sign

Arguments

:: Tensor
input

-> Tensor
output

Returns a new tensor with the signs of the elements of input

transpose

Arguments

:: Tensor
input

-> Int
dim1

-> Int
dim2

-> Tensor

Returns a tensor that is a transposed version of input. The given dimensions dim0 and dim1 are swapped.

transpose2D

Arguments

:: Tensor
input

-> Tensor
output

transpose special case for a 2D tensor

diag

Arguments

:: Tensor
input

-> Int
diagonal

-> Tensor
output

Returns a tensor with the elements of input as the diagonal. The second argument controls which diagonal to consider: If Int = 0, it is the main diagonal. If Int > 0, it is above the main diagonal. If Int < 0, it is below the main diagonal.

all

Arguments

:: Tensor
input

-> Bool
output

Returns True if all elements in the tensor are True, False otherwise.

any

Arguments

:: Tensor
input

-> Bool
output

Returns True if any elements in the tensor are True, False otherwise.

allDim

Arguments

:: Tensor
input

-> Int
dimension

-> Bool
boolean corresponding to keepdim

-> Tensor
output

Returns True if all elements in each row of the tensor in the given dimension dim are True, False otherwise. If keepdim is True, the output tensor is of the same size as input except in the dimension dim where it is of size 1. Otherwise, dim is squeezed, resulting in the output tensor having 1 fewer dimension than input.

anyDim

Arguments

:: Tensor
input

-> Int
dimension

-> Bool
boolean corresponding to keepdim

-> Tensor

Returns True if any elements in each row of the tensor in the given dimension dim are True, False otherwise. If keepdim is True, the output tensor is of the same size as input except in the dimension dim where it is of size 1. Otherwise, dim is squeezed, resulting in the output tensor having 1 fewer dimension than input.

permute

Arguments

:: Tensor
input

-> [Int]
list corresponding to ordering of dimensions to permute with

-> Tensor

Permute the dimensions of this tensor.

expand

Arguments

:: Tensor
input

-> Bool
some boolean value with unknown function

-> [Int]
the desired expanded size

-> Tensor
output

expand TODO: figure out what the boolean value does

flattenAll

Arguments

:: Tensor
input

-> Tensor
output

flattenAll

addmv :: Tensor -> Tensor -> Tensor -> Float -> Float -> Tensor

addr :: Tensor -> Tensor -> Tensor -> Float -> Float -> Tensor

allclose :: Tensor -> Tensor -> Double -> Double -> Bool -> Bool

argmin :: Tensor -> Int -> Bool -> Tensor

baddbmm :: Tensor -> Tensor -> Tensor -> Float -> Float -> Tensor

bmm :: Tensor -> Tensor -> Tensor

acos :: Tensor -> Tensor

asin :: Tensor -> Tensor

atan :: Tensor -> Tensor

dot :: Tensor -> Tensor -> Tensor

einsum :: String -> [Tensor] -> Tensor

lstsq :: Tensor -> Tensor -> (Tensor, Tensor)

mv :: Tensor -> Tensor -> Tensor

slice :: Tensor -> Int -> Int -> Int -> Int -> Tensor

sumWithDimnames :: Tensor -> [Dimname] -> Bool -> DType -> Tensor

Orphan instances

Fractional Tensor

Instance details

Methods

(/) :: Tensor -> Tensor -> Tensor

recip :: Tensor -> Tensor

fromRational :: Rational -> Tensor

Num Tensor

Instance details

Methods

(+) :: Tensor -> Tensor -> Tensor

(-) :: Tensor -> Tensor -> Tensor

(*) :: Tensor -> Tensor -> Tensor

negate :: Tensor -> Tensor

abs :: Tensor -> Tensor

signum :: Tensor -> Tensor

fromInteger :: Integer -> Tensor