Torch.Tensor

Documentation

type ATenTensor = ForeignPtr Tensor

newtype Tensor

Constructors

Unsafe ATenTensor

numel

Arguments

:: Tensor
input

-> Int
number of elements in tensor

Returns the total number of elements in the input tensor.

size

Arguments

:: Int
dimension

-> Tensor
input

-> Int

Returns the size of a given dimension of the input tensor.

shape

Arguments

:: Tensor
input

-> [Int]
list of integers representing the shape of the tensor

Returns the shape of the tensor

dim

Arguments

:: Tensor
input

-> Int
output

Returns the dimensions of the input tensor

device

Arguments

:: Tensor
input

-> Device
object representing the device

Returns the device on which the tensor is currently allocated

dtype

Arguments

:: Tensor
input

-> DType
data type of the input tensor

Returns the data type of the input tensor

toDouble :: Tensor -> Double

toInt :: Tensor -> Int

_toType

Arguments

:: DType
data type to cast input to

-> Tensor
input

-> Tensor
output

Casts the input tensor to the given data type

toType :: forall a. HasTypes a Tensor => DType -> a -> a

toDevice :: forall a. HasTypes a Tensor => Device -> a -> a

_toDevice

Arguments

:: Device
device to cast input to

-> Tensor
input

-> Tensor
output

Casts the input tensor to given device

select

Arguments

:: Int
dimension to slice along

-> Int
index in the given dimension

-> Tensor
input

-> Tensor
output

Slices the input tensor along the selected dimension at the given index.

indexSelect

Arguments

:: Int
dim

-> Tensor
indexTensor

-> Tensor
input

-> Tensor
output

Returns a new tensor which indexes the input tensor along dimension dim using the entries in index which is a LongTensor.

indexSelect'

Arguments

:: Int
dim

-> [Int]
indexList

-> Tensor
input

-> Tensor
output

sliceDim

Arguments

:: Int
dim

-> Int
start

-> Int
end

-> Int
step

-> Tensor
input

-> Tensor

Slices the input tensor along the selected dimension at the given range.

isContiguous :: Tensor -> Bool

contiguous :: Tensor -> Tensor

reshape :: [Int] -> Tensor -> Tensor

Returns a tensor with the same data and number of elements as input, but with the specified shape.

toSparse :: Tensor -> Tensor

toDense :: Tensor -> Tensor

toMKLDNN :: Tensor -> Tensor

toCPU :: Tensor -> Tensor

toCUDA :: Tensor -> Tensor

withTensorOptions :: Tensor -> TensorOptions -> Tensor

newtype RawTensorIndexList

Constructors

RawTensorIndexList (ForeignPtr (StdVector TensorIndex))

newtype RawTensorIndex

Constructors

RawTensorIndex (ForeignPtr TensorIndex)

(!) :: TensorIndex a => Tensor -> a -> Tensor

maskedFill :: (TensorIndex a, TensorLike t) => Tensor -> a -> t -> Tensor

data None

Constructors

None

data Ellipsis

Constructors

Ellipsis

newtype Slice a

Constructors

Slice a

class TensorIndex a where

Minimal complete definition

pushIndex

Methods

pushIndex :: [RawTensorIndex] -> a -> [RawTensorIndex]

toLens :: a -> Lens' Tensor Tensor

default toLens :: a -> Lens' Tensor Tensor

asValue :: TensorLike a => Tensor -> a

class TensorLike a where

Minimal complete definition

asTensor, _asValue, _dtype, _dims, _deepDims, _peekElemOff, _pokeElemOff

Methods

asTensor' :: a -> TensorOptions -> Tensor

asTensor :: a -> Tensor

_asValue :: Tensor -> a

_dtype :: DType

_dims :: a -> [Int]

_deepDims :: a -> Maybe [Int]

_peekElemOff :: Ptr () -> Int -> [Int] -> IO a

_pokeElemOff :: Ptr () -> Int -> a -> IO ()

bool_opts :: TensorOptions

uint8_opts :: TensorOptions

int64_opts :: TensorOptions

float_opts :: TensorOptions

double_opts :: TensorOptions

withTensor :: Tensor -> (Ptr () -> IO a) -> IO a

class AsTensors as where

Minimal complete definition

Nothing

Methods

toTensors :: as -> Vector Tensor

default toTensors :: (Generic as, GAsTensors (Rep as)) => as -> Vector Tensor

class GAsTensors record where

Methods

gToTensors :: record as -> Vector Tensor

Orphan instances

HasTypes Bool Bool

Instance details

Methods

types_ :: Traversal' Bool Bool

HasTypes Double Double

Instance details

Methods

types_ :: Traversal' Double Double

HasTypes Float Float

Instance details

Methods

types_ :: Traversal' Float Float

HasTypes Int Int

Instance details

Methods

types_ :: Traversal' Int Int