Torch.Typed.NN.Recurrent.GRU

Documentation

data GRULayerSpec (inputSize :: Nat) (hiddenSize :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

GRULayerSpec

data GRULayer (inputSize :: Nat) (hiddenSize :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

GRUUnidirectionalLayer :: Parameter device dtype (GRUWIShape hiddenSize inputSize) -> Parameter device dtype (GRUWHShape hiddenSize inputSize) -> Parameter device dtype (GRUBIShape hiddenSize inputSize) -> Parameter device dtype (GRUBHShape hiddenSize inputSize) -> GRULayer inputSize hiddenSize 'Unidirectional dtype device

GRUBidirectionalLayer :: Parameter device dtype (GRUWIShape hiddenSize inputSize) -> Parameter device dtype (GRUWHShape hiddenSize inputSize) -> Parameter device dtype (GRUBIShape hiddenSize inputSize) -> Parameter device dtype (GRUBHShape hiddenSize inputSize) -> Parameter device dtype (GRUWIShape hiddenSize inputSize) -> Parameter device dtype (GRUWHShape hiddenSize inputSize) -> Parameter device dtype (GRUBIShape hiddenSize inputSize) -> Parameter device dtype (GRUBHShape hiddenSize inputSize) -> GRULayer inputSize hiddenSize 'Bidirectional dtype device

data GRULayerStackSpec (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

GRULayerStackSpec

data GRULayerStack (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

GRULayer1 :: GRULayer inputSize hiddenSize directionality dtype device -> GRULayerStack inputSize hiddenSize 1 directionality dtype device

GRULayerK :: GRULayer (hiddenSize * NumberOfDirections directionality) hiddenSize directionality dtype device -> GRULayerStack inputSize hiddenSize numLayers directionality dtype device -> GRULayerStack inputSize hiddenSize (numLayers + 1) directionality dtype device

class GRULayerStackParameterized (flag :: Bool) inputSize hiddenSize numLayers directionality dtype device where

Associated Types

type GRULayerStackParameters flag inputSize hiddenSize numLayers directionality dtype device :: [Type]

Methods

gruLayerStackFlattenParameters :: Proxy flag -> GRULayerStack inputSize hiddenSize numLayers directionality dtype device -> HList (GRULayerStackParameters flag inputSize hiddenSize numLayers directionality dtype device)

gruLayerStackReplaceParameters :: Proxy flag -> GRULayerStack inputSize hiddenSize numLayers directionality dtype device -> HList (GRULayerStackParameters flag inputSize hiddenSize numLayers directionality dtype device) -> GRULayerStack inputSize hiddenSize numLayers directionality dtype device

class GRULayerStackRandomizable (flag :: Bool) inputSize hiddenSize numLayers directionality dtype device where

Methods

gruLayerStackSample :: Proxy flag -> GRULayerStackSpec inputSize hiddenSize numLayers directionality dtype device -> IO (GRULayerStack inputSize hiddenSize numLayers directionality dtype device)

newtype GRUSpec (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

GRUSpec DropoutSpec

data GRU (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

GRU

Fields

:: 1 <= numLayers

=> { gru_layer_stack :: GRULayerStack inputSize hiddenSize numLayers directionality dtype device

, gru_dropout :: Dropout

} -> GRU inputSize hiddenSize numLayers directionality dtype device

xavierUniformGRU :: forall device dtype hiddenSize featureSize. (KnownDType dtype, KnownNat hiddenSize, KnownNat featureSize, KnownDevice device, RandDTypeIsValid device dtype) => IO (Tensor device dtype '[3 * hiddenSize, featureSize])

Helper to do xavier uniform initializations on weight matrices and orthagonal initializations for the gates. (When implemented.)

data GRUWithInitSpec (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (initialization :: RNNInitialization) (dtype :: DType) (device :: (DeviceType, Nat)) where

A specification for a long, short-term memory layer.

Constructors

GRUWithZerosInitSpec :: forall inputSize hiddenSize numLayers directionality dtype device. GRUSpec inputSize hiddenSize numLayers directionality dtype device -> GRUWithInitSpec inputSize hiddenSize numLayers directionality 'ConstantInitialization dtype device
Weights drawn from Xavier-Uniform with zeros-value initialized biases and cell states.

GRUWithConstInitSpec
Weights drawn from Xavier-Uniform with zeros-value initialized biases and user-provided cell states.

Fields

:: forall inputSize hiddenSize numLayers directionality dtype device. GRUSpec inputSize hiddenSize numLayers directionality dtype device

-> Tensor device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

The initial values of the hidden state

-> GRUWithInitSpec inputSize hiddenSize numLayers directionality 'ConstantInitialization dtype device

GRUWithLearnedInitSpec
Weights drawn from Xavier-Uniform with zeros-value initialized biases and learned cell states.

Fields

:: forall inputSize hiddenSize numLayers directionality dtype device. GRUSpec inputSize hiddenSize numLayers directionality dtype device

-> Tensor device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

The initial (learnable) values of the hidden state

-> GRUWithInitSpec inputSize hiddenSize numLayers directionality 'LearnedInitialization dtype device

data GRUWithInit (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (initialization :: RNNInitialization) (dtype :: DType) (device :: (DeviceType, Nat)) where

A long, short-term memory layer with either fixed initial states for the memory cells and hidden state or learnable inital states for the memory cells and hidden state.

Constructors

GRUWithConstInit

Fields

:: forall inputSize hiddenSize numLayers directionality dtype device. { gruWithConstInit_gru :: GRU inputSize hiddenSize numLayers directionality dtype device

, gruWithConstInit_h :: Tensor device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

} -> GRUWithInit inputSize hiddenSize numLayers directionality 'ConstantInitialization dtype device

GRUWithLearnedInit

Fields

:: forall inputSize hiddenSize numLayers directionality dtype device. { gruWithLearnedInit_gru :: GRU inputSize hiddenSize numLayers directionality dtype device

, gruWithLearnedInit_h :: Parameter device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

} -> GRUWithInit inputSize hiddenSize numLayers directionality 'LearnedInitialization dtype device

gruForward :: forall shapeOrder batchSize seqLen directionality initialization numLayers inputSize outputSize hiddenSize inputShape outputShape hcShape parameters tensorParameters dtype device. (KnownNat (NumberOfDirections directionality), KnownNat numLayers, KnownNat batchSize, KnownNat hiddenSize, KnownRNNShapeOrder shapeOrder, KnownRNNDirectionality directionality, outputSize ~ (hiddenSize * NumberOfDirections directionality), inputShape ~ RNNShape shapeOrder seqLen batchSize inputSize, outputShape ~ RNNShape shapeOrder seqLen batchSize outputSize, hcShape ~ '[numLayers * NumberOfDirections directionality, batchSize, hiddenSize], parameters ~ Parameters (GRU inputSize hiddenSize numLayers directionality dtype device), Parameterized (GRU inputSize hiddenSize numLayers directionality dtype device), tensorParameters ~ GRUR inputSize hiddenSize numLayers directionality dtype device, Castable (HList tensorParameters) [ATenTensor], HMap' ToDependent parameters tensorParameters) => Bool -> GRUWithInit inputSize hiddenSize numLayers directionality initialization dtype device -> Tensor device dtype inputShape -> (Tensor device dtype outputShape, Tensor device dtype hcShape)

gruForwardWithDropout :: forall shapeOrder batchSize seqLen directionality initialization numLayers inputSize outputSize hiddenSize inputShape outputShape hcShape parameters tensorParameters dtype device. (KnownNat (NumberOfDirections directionality), KnownNat numLayers, KnownNat batchSize, KnownNat hiddenSize, KnownRNNShapeOrder shapeOrder, KnownRNNDirectionality directionality, outputSize ~ (hiddenSize * NumberOfDirections directionality), inputShape ~ RNNShape shapeOrder seqLen batchSize inputSize, outputShape ~ RNNShape shapeOrder seqLen batchSize outputSize, hcShape ~ '[numLayers * NumberOfDirections directionality, batchSize, hiddenSize], parameters ~ Parameters (GRU inputSize hiddenSize numLayers directionality dtype device), Parameterized (GRU inputSize hiddenSize numLayers directionality dtype device), tensorParameters ~ GRUR inputSize hiddenSize numLayers directionality dtype device, Castable (HList tensorParameters) [ATenTensor], HMap' ToDependent parameters tensorParameters) => GRUWithInit inputSize hiddenSize numLayers directionality initialization dtype device -> Tensor device dtype inputShape -> (Tensor device dtype outputShape, Tensor device dtype hcShape)

Forward propagate the GRU module (without applying dropout on the outputs of each layer).

>>> input :: CPUTensor 'D.Float '[5,16,10] <- randn
>>> spec = GRUWithZerosInitSpec @10 @30 @3 @'Bidirectional @'D.Float @'( 'D.CPU, 0) (GRUSpec (DropoutSpec 0.5))
>>> model <- A.sample spec
>>> :t gruForwardWithoutDropout @'BatchFirst model input
gruForwardWithoutDropout @'BatchFirst model input
  :: (Tensor '( 'D.CPU, 0) 'D.Float '[5, 16, 60],
      Tensor '( 'D.CPU, 0) 'D.Float '[6, 5, 30])
>>> (a,b) = gruForwardWithoutDropout @'BatchFirst model input
>>> ((dtype a, shape a), (dtype b, shape b))
((Float,[5,16,60]),(Float,[6,5,30]))

Forward propagate the GRU module and apply dropout on the outputs of each layer.

>>> input :: CPUTensor 'D.Float '[5,16,10] <- randn
>>> spec = GRUWithZerosInitSpec @10 @30 @3 @'Bidirectional @'D.Float @'( 'D.CPU, 0) (GRUSpec (DropoutSpec 0.5))
>>> model <- A.sample spec
>>> :t gruForwardWithDropout @'BatchFirst model input
gruForwardWithDropout @'BatchFirst model input
  :: (Tensor '( 'D.CPU, 0) 'D.Float '[5, 16, 60],
      Tensor '( 'D.CPU, 0) 'D.Float '[6, 5, 30])
>>> (a,b) = gruForwardWithDropout @'BatchFirst model input
>>> ((dtype a, shape a), (dtype b, shape b))
((Float,[5,16,60]),(Float,[6,5,30]))

gruForwardWithoutDropout :: forall shapeOrder batchSize seqLen directionality initialization numLayers inputSize outputSize hiddenSize inputShape outputShape hcShape parameters tensorParameters dtype device. (KnownNat (NumberOfDirections directionality), KnownNat numLayers, KnownNat batchSize, KnownNat hiddenSize, KnownRNNShapeOrder shapeOrder, KnownRNNDirectionality directionality, outputSize ~ (hiddenSize * NumberOfDirections directionality), inputShape ~ RNNShape shapeOrder seqLen batchSize inputSize, outputShape ~ RNNShape shapeOrder seqLen batchSize outputSize, hcShape ~ '[numLayers * NumberOfDirections directionality, batchSize, hiddenSize], parameters ~ Parameters (GRU inputSize hiddenSize numLayers directionality dtype device), Parameterized (GRU inputSize hiddenSize numLayers directionality dtype device), tensorParameters ~ GRUR inputSize hiddenSize numLayers directionality dtype device, Castable (HList tensorParameters) [ATenTensor], HMap' ToDependent parameters tensorParameters) => GRUWithInit inputSize hiddenSize numLayers directionality initialization dtype device -> Tensor device dtype inputShape -> (Tensor device dtype outputShape, Tensor device dtype hcShape)

Forward propagate the GRU module and apply dropout on the outputs of each layer.

>>> input :: CPUTensor 'D.Float '[5,16,10] <- randn
>>> spec = GRUWithZerosInitSpec @10 @30 @3 @'Bidirectional @'D.Float @'( 'D.CPU, 0) (GRUSpec (DropoutSpec 0.5))
>>> model <- A.sample spec
>>> :t gruForwardWithDropout @'BatchFirst model input
gruForwardWithDropout @'BatchFirst model input
  :: (Tensor '( 'D.CPU, 0) 'D.Float '[5, 16, 60],
      Tensor '( 'D.CPU, 0) 'D.Float '[6, 5, 30])
>>> (a,b) = gruForwardWithDropout @'BatchFirst model input
>>> ((dtype a, shape a), (dtype b, shape b))
((Float,[5,16,60]),(Float,[6,5,30]))