Torch.Typed.NN.Recurrent.LSTM

Documentation

data LSTMLayerSpec (inputSize :: Nat) (hiddenSize :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

LSTMLayerSpec

data LSTMLayer (inputSize :: Nat) (hiddenSize :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

LSTMUnidirectionalLayer :: Parameter device dtype (LSTMWIShape hiddenSize inputSize) -> Parameter device dtype (LSTMWHShape hiddenSize inputSize) -> Parameter device dtype (LSTMBIShape hiddenSize inputSize) -> Parameter device dtype (LSTMBHShape hiddenSize inputSize) -> LSTMLayer inputSize hiddenSize 'Unidirectional dtype device

LSTMBidirectionalLayer :: Parameter device dtype (LSTMWIShape hiddenSize inputSize) -> Parameter device dtype (LSTMWHShape hiddenSize inputSize) -> Parameter device dtype (LSTMBIShape hiddenSize inputSize) -> Parameter device dtype (LSTMBHShape hiddenSize inputSize) -> Parameter device dtype (LSTMWIShape hiddenSize inputSize) -> Parameter device dtype (LSTMWHShape hiddenSize inputSize) -> Parameter device dtype (LSTMBIShape hiddenSize inputSize) -> Parameter device dtype (LSTMBHShape hiddenSize inputSize) -> LSTMLayer inputSize hiddenSize 'Bidirectional dtype device

data LSTMLayerStackSpec (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

LSTMLayerStackSpec

data LSTMLayerStack (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

LSTMLayer1 :: LSTMLayer inputSize hiddenSize directionality dtype device -> LSTMLayerStack inputSize hiddenSize 1 directionality dtype device

LSTMLayerK :: LSTMLayer (hiddenSize * NumberOfDirections directionality) hiddenSize directionality dtype device -> LSTMLayerStack inputSize hiddenSize numLayers directionality dtype device -> LSTMLayerStack inputSize hiddenSize (numLayers + 1) directionality dtype device

class LSTMLayerStackParameterized (flag :: Bool) inputSize hiddenSize numLayers directionality dtype device where

Associated Types

type LSTMLayerStackParameters flag inputSize hiddenSize numLayers directionality dtype device :: [Type]

Methods

lstmLayerStackFlattenParameters :: Proxy flag -> LSTMLayerStack inputSize hiddenSize numLayers directionality dtype device -> HList (LSTMLayerStackParameters flag inputSize hiddenSize numLayers directionality dtype device)

lstmLayerStackReplaceParameters :: Proxy flag -> LSTMLayerStack inputSize hiddenSize numLayers directionality dtype device -> HList (LSTMLayerStackParameters flag inputSize hiddenSize numLayers directionality dtype device) -> LSTMLayerStack inputSize hiddenSize numLayers directionality dtype device

class LSTMLayerStackRandomizable (flag :: Bool) inputSize hiddenSize numLayers directionality dtype device where

Methods

lstmLayerStackSample :: Proxy flag -> LSTMLayerStackSpec inputSize hiddenSize numLayers directionality dtype device -> IO (LSTMLayerStack inputSize hiddenSize numLayers directionality dtype device)

newtype LSTMSpec (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat))

Constructors

LSTMSpec DropoutSpec

data LSTM (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (dtype :: DType) (device :: (DeviceType, Nat)) where

Constructors

LSTM

Fields

:: 1 <= numLayers

=> { lstm_layer_stack :: LSTMLayerStack inputSize hiddenSize numLayers directionality dtype device

, lstm_dropout :: Dropout

} -> LSTM inputSize hiddenSize numLayers directionality dtype device

xavierUniformLSTM :: forall device dtype hiddenSize featureSize. (KnownDType dtype, KnownNat hiddenSize, KnownNat featureSize, KnownDevice device, RandDTypeIsValid device dtype) => IO (Tensor device dtype '[4 * hiddenSize, featureSize])

Helper to do xavier uniform initializations on weight matrices and orthagonal initializations for the gates. (When implemented.)

data LSTMWithInitSpec (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (initialization :: RNNInitialization) (dtype :: DType) (device :: (DeviceType, Nat)) where

A specification for a long, short-term memory layer.

Constructors

LSTMWithZerosInitSpec :: forall inputSize hiddenSize numLayers directionality dtype device. LSTMSpec inputSize hiddenSize numLayers directionality dtype device -> LSTMWithInitSpec inputSize hiddenSize numLayers directionality 'ConstantInitialization dtype device
Weights drawn from Xavier-Uniform with zeros-value initialized biases and cell states.

LSTMWithConstInitSpec
Weights drawn from Xavier-Uniform with zeros-value initialized biases and user-provided cell states.

Fields

:: forall inputSize hiddenSize numLayers directionality dtype device. LSTMSpec inputSize hiddenSize numLayers directionality dtype device

-> Tensor device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

The initial values of the memory cell

-> Tensor device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

The initial values of the hidden state

-> LSTMWithInitSpec inputSize hiddenSize numLayers directionality 'ConstantInitialization dtype device

LSTMWithLearnedInitSpec
Weights drawn from Xavier-Uniform with zeros-value initialized biases and learned cell states.

Fields

:: forall inputSize hiddenSize numLayers directionality dtype device. LSTMSpec inputSize hiddenSize numLayers directionality dtype device

-> Tensor device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

The initial (learnable) values of the memory cell

-> Tensor device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

The initial (learnable) values of the hidden state

-> LSTMWithInitSpec inputSize hiddenSize numLayers directionality 'LearnedInitialization dtype device

data LSTMWithInit (inputSize :: Nat) (hiddenSize :: Nat) (numLayers :: Nat) (directionality :: RNNDirectionality) (initialization :: RNNInitialization) (dtype :: DType) (device :: (DeviceType, Nat)) where

A long, short-term memory layer with either fixed initial states for the memory cells and hidden state or learnable inital states for the memory cells and hidden state.

Constructors

LSTMWithConstInit

Fields

:: forall inputSize hiddenSize numLayers directionality dtype device. { lstmWithConstInit_lstm :: LSTM inputSize hiddenSize numLayers directionality dtype device

, lstmWithConstInit_c :: Tensor device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

, lstmWithConstInit_h :: Tensor device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

} -> LSTMWithInit inputSize hiddenSize numLayers directionality 'ConstantInitialization dtype device

LSTMWithLearnedInit

Fields

:: forall inputSize hiddenSize numLayers directionality dtype device. { lstmWithLearnedInit_lstm :: LSTM inputSize hiddenSize numLayers directionality dtype device

, lstmWithLearnedInit_c :: Parameter device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

, lstmWithLearnedInit_h :: Parameter device dtype '[numLayers * NumberOfDirections directionality, hiddenSize]

} -> LSTMWithInit inputSize hiddenSize numLayers directionality 'LearnedInitialization dtype device

lstmForward :: forall shapeOrder batchSize seqLen directionality initialization numLayers inputSize outputSize hiddenSize inputShape outputShape hxShape parameters tensorParameters dtype device. (KnownNat (NumberOfDirections directionality), KnownNat numLayers, KnownNat batchSize, KnownNat hiddenSize, KnownRNNShapeOrder shapeOrder, KnownRNNDirectionality directionality, outputSize ~ (hiddenSize * NumberOfDirections directionality), inputShape ~ RNNShape shapeOrder seqLen batchSize inputSize, outputShape ~ RNNShape shapeOrder seqLen batchSize outputSize, hxShape ~ '[numLayers * NumberOfDirections directionality, batchSize, hiddenSize], parameters ~ Parameters (LSTM inputSize hiddenSize numLayers directionality dtype device), Parameterized (LSTM inputSize hiddenSize numLayers directionality dtype device), tensorParameters ~ LSTMR inputSize hiddenSize numLayers directionality dtype device, Castable (HList tensorParameters) [ATenTensor], HMap' ToDependent parameters tensorParameters) => Bool -> LSTMWithInit inputSize hiddenSize numLayers directionality initialization dtype device -> Tensor device dtype inputShape -> (Tensor device dtype outputShape, Tensor device dtype hxShape, Tensor device dtype hxShape)

lstmForwardWithDropout :: forall shapeOrder batchSize seqLen directionality initialization numLayers inputSize outputSize hiddenSize inputShape outputShape hxShape parameters tensorParameters dtype device. (KnownNat (NumberOfDirections directionality), KnownNat numLayers, KnownNat batchSize, KnownNat hiddenSize, KnownRNNShapeOrder shapeOrder, KnownRNNDirectionality directionality, outputSize ~ (hiddenSize * NumberOfDirections directionality), inputShape ~ RNNShape shapeOrder seqLen batchSize inputSize, outputShape ~ RNNShape shapeOrder seqLen batchSize outputSize, hxShape ~ '[numLayers * NumberOfDirections directionality, batchSize, hiddenSize], parameters ~ Parameters (LSTM inputSize hiddenSize numLayers directionality dtype device), Parameterized (LSTM inputSize hiddenSize numLayers directionality dtype device), tensorParameters ~ LSTMR inputSize hiddenSize numLayers directionality dtype device, Castable (HList tensorParameters) [ATenTensor], HMap' ToDependent parameters tensorParameters) => LSTMWithInit inputSize hiddenSize numLayers directionality initialization dtype device -> Tensor device dtype inputShape -> (Tensor device dtype outputShape, Tensor device dtype hxShape, Tensor device dtype hxShape)

Forward propagate the LSTM module (without applying dropout on the outputs of each layer).

>>> input :: CPUTensor 'D.Float '[5,16,10] <- randn
>>> spec = LSTMWithZerosInitSpec @10 @30 @3 @'Bidirectional @'D.Float @'( 'D.CPU, 0) (LSTMSpec (DropoutSpec 0.5))
>>> model <- A.sample spec
>>> :t lstmForwardWithoutDropout @'BatchFirst model input
lstmForwardWithoutDropout @'BatchFirst model input
  :: (Tensor '( 'D.CPU, 0) 'D.Float '[5, 16, 60],
      Tensor '( 'D.CPU, 0) 'D.Float '[6, 5, 30],
      Tensor '( 'D.CPU, 0) 'D.Float '[6, 5, 30])
>>> (a,b,c) = lstmForwardWithoutDropout @'BatchFirst model input
>>> ((dtype a, shape a), (dtype b, shape b), (dtype c, shape c))
((Float,[5,16,60]),(Float,[6,5,30]),(Float,[6,5,30]))

Forward propagate the LSTM module and apply dropout on the outputs of each layer.

>>> input :: CPUTensor 'D.Float '[5,16,10] <- randn
>>> spec = LSTMWithZerosInitSpec @10 @30 @3 @'Bidirectional @'D.Float @'( 'D.CPU, 0) (LSTMSpec (DropoutSpec 0.5))
>>> model <- A.sample spec
>>> :t lstmForwardWithDropout @'BatchFirst model input
lstmForwardWithDropout @'BatchFirst model input
  :: (Tensor '( 'D.CPU, 0) 'D.Float '[5, 16, 60],
      Tensor '( 'D.CPU, 0) 'D.Float '[6, 5, 30],
      Tensor '( 'D.CPU, 0) 'D.Float '[6, 5, 30])
>>> (a,b,c) = lstmForwardWithDropout @'BatchFirst model input
>>> ((dtype a, shape a), (dtype b, shape b), (dtype c, shape c))
((Float,[5,16,60]),(Float,[6,5,30]),(Float,[6,5,30]))

lstmForwardWithoutDropout :: forall shapeOrder batchSize seqLen directionality initialization numLayers inputSize outputSize hiddenSize inputShape outputShape hxShape parameters tensorParameters dtype device. (KnownNat (NumberOfDirections directionality), KnownNat numLayers, KnownNat batchSize, KnownNat hiddenSize, KnownRNNShapeOrder shapeOrder, KnownRNNDirectionality directionality, outputSize ~ (hiddenSize * NumberOfDirections directionality), inputShape ~ RNNShape shapeOrder seqLen batchSize inputSize, outputShape ~ RNNShape shapeOrder seqLen batchSize outputSize, hxShape ~ '[numLayers * NumberOfDirections directionality, batchSize, hiddenSize], parameters ~ Parameters (LSTM inputSize hiddenSize numLayers directionality dtype device), Parameterized (LSTM inputSize hiddenSize numLayers directionality dtype device), tensorParameters ~ LSTMR inputSize hiddenSize numLayers directionality dtype device, Castable (HList tensorParameters) [ATenTensor], HMap' ToDependent parameters tensorParameters) => LSTMWithInit inputSize hiddenSize numLayers directionality initialization dtype device -> Tensor device dtype inputShape -> (Tensor device dtype outputShape, Tensor device dtype hxShape, Tensor device dtype hxShape)

Forward propagate the LSTM module and apply dropout on the outputs of each layer.

>>> input :: CPUTensor 'D.Float '[5,16,10] <- randn
>>> spec = LSTMWithZerosInitSpec @10 @30 @3 @'Bidirectional @'D.Float @'( 'D.CPU, 0) (LSTMSpec (DropoutSpec 0.5))
>>> model <- A.sample spec
>>> :t lstmForwardWithDropout @'BatchFirst model input
lstmForwardWithDropout @'BatchFirst model input
  :: (Tensor '( 'D.CPU, 0) 'D.Float '[5, 16, 60],
      Tensor '( 'D.CPU, 0) 'D.Float '[6, 5, 30],
      Tensor '( 'D.CPU, 0) 'D.Float '[6, 5, 30])
>>> (a,b,c) = lstmForwardWithDropout @'BatchFirst model input
>>> ((dtype a, shape a), (dtype b, shape b), (dtype c, shape c))
((Float,[5,16,60]),(Float,[6,5,30]),(Float,[6,5,30]))